+++ custom_components/ge_home/entities/__init__.py
@@ -5,3 +5,4 @@ from .oven import *
 from .water_filter import *
 from .advantium import *
 from .ac import *
+from .hood import *