+++ custom_components/ge_home/sensor.py
@@ -2,13 +2,24 @@
 import async_timeout
 import logging
 from typing import Callable
+import voluptuous as vol
+from datetime import timedelta
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers import entity_platform
 
-from .const import DOMAIN
+from .const import (
+    DOMAIN, 
+    SERVICE_SET_TIMER, 
+    SERVICE_CLEAR_TIMER, 
+    SERVICE_SET_INT_VALUE
+)
 from .entities import GeErdSensor
 from .update_coordinator import GeHomeUpdateCoordinator
+
+ATTR_DURATION = "duration"
+ATTR_VALUE = "value"
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -16,6 +27,9 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry, asyn
     """GE Home sensors."""
     _LOGGER.debug('Adding GE Home sensors')
     coordinator: GeHomeUpdateCoordinator = hass.data[DOMAIN][config_entry.entry_id]
+    
+    # Get the platform
+    platform = entity_platform.async_get_current_platform()
 
     # This should be a NOP, but let's be safe
     with async_timeout.timeout(20):
@@ -32,3 +46,33 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry, asyn
     ]
     _LOGGER.debug(f'Found {len(entities):d} sensors')
     async_add_entities(entities)
+
+    # register set_timer entity service
+    platform.async_register_entity_service(
+    SERVICE_SET_TIMER,
+    {
+        vol.Required(ATTR_DURATION): vol.All(
+            vol.Coerce(int), vol.Range(min=1, max=360)
+        )
+    },
+    set_timer)
+
+    # register clear_timer entity service
+    platform.async_register_entity_service(SERVICE_CLEAR_TIMER, {}, 'clear_timer')
+
+    # register set_value entity service
+    platform.async_register_entity_service(
+    SERVICE_SET_INT_VALUE,
+    {
+        vol.Required(ATTR_VALUE): vol.All(
+            vol.Coerce(int), vol.Range(min=0)
+        )
+    },
+    set_int_value)    
+
+async def set_timer(entity, service_call):
+    ts = timedelta(minutes=int(service_call.data['duration']))
+    await entity.set_timer(ts)
+
+async def set_int_value(entity, service_call):
+    await entity.set_value(int(service_call.data['value']))