+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Initial support for oven hoods (@digitalbites)
 - Added extended mode support for ovens
 - Added logic to prevent multiple configurations of the same GE account
+- Fixed device info when serial not present (@Xe138)
 
 ## 0.4.3